회원 시큐리티 로직 개선 (#388)
* refactor: 유저 정보를 가져오는 CustomUserDetails 클래스 리팩토링

* chore: config -> Member 패키지로 이동

---------
