+++ src/main/java/balancetalk/member/application/MyUserDetailService.java
@@ -1,17 +1,15 @@
-package balancetalk.global.config;
+package balancetalk.member.application;
 
+import balancetalk.member.domain.CustomUserDetails;
 import balancetalk.global.exception.BalanceTalkException;
 import balancetalk.global.exception.ErrorCode;
 import balancetalk.member.domain.Member;
 import balancetalk.member.domain.MemberRepository;
 import lombok.RequiredArgsConstructor;
-import org.springframework.security.core.GrantedAuthority;
 import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.security.core.userdetails.UserDetailsService;
 import org.springframework.security.core.userdetails.UsernameNotFoundException;
 import org.springframework.stereotype.Service;
-import java.util.Collections;
-import java.util.List;
 
 @Service
 @RequiredArgsConstructor
@@ -23,15 +21,6 @@ public class MyUserDetailService implements UserDetailsService {
     public UserDetails loadUserByUsername(String username) throws UsernameNotFoundException {
         Member member = memberRepository.findByEmail(username)
                 .orElseThrow(() -> new BalanceTalkException(ErrorCode.NOT_FOUND_MEMBER));
-
-        List<GrantedAuthority> authorities = Collections.emptyList();
-
-        CustomUserDetails userDetails = new CustomUserDetails(
-                member.getEmail(),
-                member.getPassword(),
-                authorities,
-                member.getId() // memberId를 추가
-        );
-        return userDetails;
+        return new CustomUserDetails(member);
     }
 }
